+++ native/src/JNICallbackAdapter.h
@@ -33,7 +33,7 @@ private:
 	
 	jobject bufferReference;
 	
-	jint sequence;
+	jlong sequence;
 	
 	// Is this a read operation
 	short isRead;
@@ -50,7 +50,7 @@ private:
 	
 public:
 	// _ob must be a global Reference (use createGloblReferente before calling the constructor)
-	JNICallbackAdapter(AIOController * _controller, jint sequence, jobject _callback, jobject _fileController, jobject _bufferReference, short _isRead);
+	JNICallbackAdapter(AIOController * _controller, jlong sequence, jobject _callback, jobject _fileController, jobject _bufferReference, short _isRead);
 	virtual ~JNICallbackAdapter();
 
 	void done(THREAD_CONTEXT threadContext);
